+++ src/common/api/profiles.api.ts
@@ -2,6 +2,7 @@ import {
   PROFILE_API_ENDPOINT,
   PROFILE_METADATA_API_ENDPOINT,
   PROFILE_PREFERRED_API_ENDPOINT,
+  PROFILE_SETTINGS_API_ENDPOINT,
 } from '@common/constants/api.constants';
 import baseApi from './base.api';
 
@@ -46,6 +47,27 @@ export const deletePreferredProfile = (resourceType: ResourceTypeURL) => {
     url,
     requestParams: {
       method: 'DELETE',
+    },
+  });
+};
+
+export const fetchProfileSettings = (profileId: string | number) =>
+  baseApi.getJson({
+    url: `${PROFILE_SETTINGS_API_ENDPOINT}/${profileId}`,
+  }) as Promise<ProfileSettings>;
+
+export const saveProfileSettings = (profileId: string | number, settings: ProfileSettings) => {
+  const url = `${PROFILE_SETTINGS_API_ENDPOINT}/${profileId}`;
+  const body = JSON.stringify(settings);
+
+  return baseApi.request({
+    url,
+    requestParams: {
+      method: 'POST',
+      body,
+      headers: {
+        'content-type': 'application/json',
+      },
     },
   });
 };